TL;DW

+++ src/main/java/com/zzzz/dao/CourseDao.java
@@ -6,7 +6,7 @@ import com.zzzz.vo.CourseDetail;
 import java.sql.SQLException;
 
 public interface CourseDao {
-    long insert(Course course) throws SQLException;
+    int insert(Course course) throws SQLException;
     boolean checkExistenceById(long courseId) throws SQLException;
     Course getById(long courseId) throws SQLException;
     CourseDetail getVoById(long courseId) throws SQLException;
